recipe: add TouchDetector components, enhance reading
Adds a `strict` flag to reading the recipe (by default activated) that
will raise upon encountering attributes in XML tags that are not
allowed.

Also add a more recent version of the recipe to test against, and allows
to dump the recipe into two different files, as is customary in our
toolchain(s).

Change-Id: Ibbcd51ef6ae09aead8a0e4c9fa75d1be0366d76f