refactor: shift more from Spark to Pandas
Contains computations that need to be broadcasted and/or are pretty
small to the master node and avoids some overhead.  On larger clusters,
this will save about 5 minutes of runtime when instantiating the
filters.

Also consolidate the column generation and use a single mapping in
Pandas to generate properties. This shifts the code both further towards
the new API and reduces the number of calls into Python.

Change-Id: I4f0809c927e5a9555b25eed141a9e8d152fedd16