FUNCZ-318 Remove mandatory recipe part
Defaults to a bogus recipe that will raise an exception when accessed. This closes FUNCZ-318, allowing filters that do not require a recipe to run without any fake recipe injected just to satisfy the application requirements.

Also delays touch loading to happen after filter instantiation, to ensure that errors due to a missing recipe are raised before any expensive computation.