FUNCZ-333 Detect if section type shift is needed.
Historically, TouchDetector saves the section type as 0-based index.
These used to be converted as-is, contrary to the MorphIO convention of
1-based type indices.

Recently, Spykfunc expects inputs to be closer to the SONATA standard in
both naming and conventions. Thus input files that were recently
converted have their section type shifted to be 0-based internally, with
a default of not shifting any non-converted Parquet input files. This
breaks projection inputs, as they adhere to the SONATA specification and
are handled with 1-based values inside Spykfunc.

Add some heuristics that if we can determine which section types are
involved and thus shift automatically if no Parquet conversion
information is present.