FUNCZ-309 Remove repeated joins of nodes to edges
Reading in the node dataframe, reducing to a single partition and broadcasting it breaks down for very large node dataframes. This enabled me to run on the 75 mio node circuit without tuning any memory parameters.

Thus we should read it in, join it to the edge dataframe once and proceed as before.  This, as a side effect, also reduces our code complexity and the amount of SQL queries needing to be done.

As more data is saved for each checkpoint when processing, space requirements increase by roughly 10% (anecdotal evidence from local tests with the 1k/2k test circuits). The runtime on the S1 circuit is very similar to before, 26 minutes rather than 25.